OPENEUROPA-2043: Create only one subscriber and update documentation.
