+++ modules/oe_webtools_cookie_consent/oe_webtools_cookie_consent.services.yml
@@ -1,11 +1,6 @@
 services:
-  oe_webtools_cookie_consent.config_banner_popup:
-    class: Drupal\oe_webtools_cookie_consent\EventSubscriber\ConfigBannerPopupEventSubscriber
-    arguments: ['@config.factory']
-    tags:
-      - { name: event_subscriber }
-  oe_webtools_cookie_consent.config_video_popup:
-    class: Drupal\oe_webtools_cookie_consent\EventSubscriber\ConfigVideoPopupEventSubscriber
+  oe_webtools_cookie_consent.event_subscriber:
+    class: Drupal\oe_webtools_cookie_consent\EventSubscriber\WebtoolsCookieConsentEventSubscriber
     arguments: ['@config.factory']
     tags:
       - { name: event_subscriber }
