+++ modules/oe_webtools_cookie_consent/src/EventSubscriber/WebtoolsCookieConsentEventSubscriber.php
@@ -5,13 +5,14 @@ declare(strict_types = 1);
 namespace Drupal\oe_webtools_cookie_consent\EventSubscriber;
 
 use Drupal\Core\Config\ConfigFactoryInterface;
+use Drupal\oe_webtools_cookie_consent\Event\ConfigBannerPopupEvent;
 use Drupal\oe_webtools_cookie_consent\Event\ConfigVideoPopupEvent;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 /**
  * Subscribes to the event fired when visitor data is collected for CCK.
  */
-class ConfigVideoPopupEventSubscriber implements EventSubscriberInterface {
+class WebtoolsCookieConsentEventSubscriber implements EventSubscriberInterface {
 
   /**
    * The config factory.
@@ -21,13 +22,27 @@ class ConfigVideoPopupEventSubscriber implements EventSubscriberInterface {
   protected $configFactory;
 
   /**
-   * Constructs an ConfigBannerPopupEventSubscriber.
+   * Constructs an WebtoolsCookieConsentEventSubscriber.
    *
    * @param \Drupal\Core\Config\ConfigFactoryInterface $configFactory
    *   The configuration object factory.
    */
   public function __construct(ConfigFactoryInterface $configFactory) {
     $this->configFactory = $configFactory;
+  }
+
+  /**
+   * Kernel request event handler.
+   *
+   * @param \Drupal\oe_webtools_cookie_consent\Event\ConfigBannerPopupEvent $event
+   *   Response event.
+   */
+  public function onSetBannerPopup(ConfigBannerPopupEvent $event): void {
+    $config = $this->configFactory->get(ConfigBannerPopupEvent::CONFIG_NAME);
+    $event->addCacheableDependency($config);
+
+    $config_data = $config->get(ConfigBannerPopupEvent::BANNER_POPUP);
+    $event->setBannerPopup((boolean) $config_data);
   }
 
   /**
@@ -49,6 +64,7 @@ class ConfigVideoPopupEventSubscriber implements EventSubscriberInterface {
    */
   public static function getSubscribedEvents(): array {
     // Subscribing to listening to the Cookie Consent event.
+    $events[ConfigBannerPopupEvent::NAME][] = ['onSetBannerPopup'];
     $events[ConfigVideoPopupEvent::NAME][] = ['onSetVideoPopup'];
 
     return $events;
